pyproject.toml
[project]
name = "demo-agent"
version = "0.1.0"
description = "Municipal data research agent - FCI Hackathon Demo"
requires-python = ">=3.10"
dependencies = [
    "deepagents",
    "langchain>=1.0",
    "langchain-core>=1.0",
    "langchain-anthropic",
    "langsmith>=0.3.0",
    "langgraph>=1.0",
    "tavily-python",
    "python-dotenv",
]
